remove SignHandler
